+++ odbc/src/api/get_info_tests.rs
@@ -98,12 +98,12 @@ macro_rules! test_get_info_expect_u32_sql_all {
 unsafe fn modify_string_value(value_ptr: Pointer, out_length: usize) -> String {
     input_text_to_string_w(
         value_ptr as *const _,
-        (out_length / std::mem::size_of::<WideChar>()) as i32,
+        (out_length / std::mem::size_of::<WideChar>()) as isize,
     )
 }
 
 unsafe fn modify_string_value_from_runes(value_ptr: Pointer, out_length: usize) -> String {
-    input_text_to_string_w(value_ptr as *const _, out_length as i32)
+    input_text_to_string_w(value_ptr as *const _, out_length as isize)
 }
 
 unsafe fn modify_u32_value(value_ptr: Pointer, _: usize) -> u32 {
